Renamed folder for rentals query related classes
