+++ src/VacationRental.Api/Controllers/CalendarController.cs
@@ -5,7 +5,7 @@ using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using VacationRental.Api.Models;
 using VacationRental.Application.Common.Models;
-using VacationRental.Application.Rentals.Queries.GetRentals;
+using VacationRental.Application.Rentals.Queries.GetRentalById;
 
 namespace VacationRental.Api.Controllers
 {
