+++ src/VacationRental.Application/Rentals/Queries/GetRentalById/GetRentalByIdQuery.cs
@@ -3,7 +3,7 @@ using System.Text;
 using MediatR;
 using VacationRental.Application.Common.Models;
 
-namespace VacationRental.Application.Rentals.Queries.GetRentals
+namespace VacationRental.Application.Rentals.Queries.GetRentalById
 {
     public class GetRentalByIdQuery : IRequest<RentalViewModel>
     {
